refactor: unify corecrypto conversation info impls in a common module
# Conflicts:
#	crypto-ffi/src/generic/mod.rs
#	crypto-ffi/src/wasm/mod.rs
